Initial version of ASP.NET vNext addin.
Adds two menu options to File - Open:

1) ASP.NET vNext Project
2) ASP.NET vNext Solution Directory

The project menu option can be used to open a project.json file.
The solution directory menu option can be used to point to a directory
where projects can exist in subdirectories.

The klr.exe is hosted and the references in the Projects window is
updated based on information returned from klr.exe